Change window split and kill buffer key bindings.
Now when we split window with "SPC w -" or "SPC w /", the new split window will
be focused and switch to scratch buffer automatically, then we can choose to
display a new file or a existing buffer in this new split window.

Function kill-buffer-and-window is not perfect, but it is the good enough one if
we don't have too many time to create a perfect one, it will kill the buffer and
its window, it will meet problem when we only have one sole ordinary window, but
actually only a message "Attempt to delete minibuffer or sole ordinary window"
appeared, the buffer will be killed correctly, so it can meet our requirement.