Change Terminal Cursor Shape & Color Dynamically to Match Evil State
Change cursor shape and color using xterm DECSCUSR and OSC 12 escape sequences
when emacs or emacsclient is running in terminal. When the emacs frame close the
cursor shape need to be restored, because xterm DECSCUSR didn't clearly define
the cursor shape reset code(even lots of terminals use 0 as the reset code), so
the cursor shape and color will be restored to appointed values, and the default
cursor shape and color of terminal applications also should be configured to
them.

Reference:
- https://invisible-island.net/xterm/ctlseqs/ctlseqs.html
- https://github.com/amosbird/evil-terminal-cursor-changer
- https://github.com/kiennq/evil-terminal-cursor-changer